requirements.txt
numpy==1.24.3
streamlit==1.32.1
Pillow==10.2.0
keras==2.15.0
tensorflow==2.13.0
